requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0.0
datasets>=2.16.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
joblib>=1.3.0
transformers>=4.38.0
torch>=2.1.0
